requirements.txt
# --- For-Bot-Working --------- #
pyrogram
TgCrypto
pyromod==1.5
# --- For-Database ------------ #
pymongo
dnspython
# --- For-Web-Response ------- #
aiohttp
requests
bs4
# --- For-Jio-DL ------- #
xmltodict
requests
pyjwt
protobuf==3.20.
pyperclip
#additionally
pycryptodomex
